좌석 구매 후 티켓예약, 좌석예약, 포인트 적립,사용 변경함
좌석 구매 후 티켓예약, 좌석예약, 포인트 적립,사용 변경함

+++ src/main/java/com/example/vo/Movie.java
@@ -16,6 +16,6 @@ public class Movie {
 	private int no;
 	private int movieRatingNo;
 	private String movieName;
-	private long movieAudienceTotalNumber;
+	private int movieAudienceTotalNumber;
 	private int likeCount;
 }
